HW 17 updated
The error message you referenced occured because the song already existed in the playlist. It does not trigger the message that the song was added to the playlist, so it fails. I tried to automate it, but it keeps telling me the element is not interactable, so I gave up. Also, are all the other methods like saerching for a song, showing results, etc. supposed to be in BaseTest?